requirements.txt
flask
python-dotenv
pandas
openai
langchain-openai
twilio
